Update how config is handled
Well turns out that gitlab-shell does not actually pass custom
environment settings that the rest of gitlab gets through gitlab.rb, so
for now I've removed any customization of the config file.

Updated the config file load to error out effectively if the file is
unreadable or missing it will report a valid error.

Will submit an issue with gitlab-shell to support adding custom
settings.