Add ability for out-of-band patching
This change adds the ability to handle a regex and use the MR API to
check the MR description for specific words/phrases and allow a merge to
occur. This could be useful in the event a security patch or bug fix
needs to occur outside the normal promotion workflow.

Also adds documentation to the README and config file