Fix Project Logic
If there are a bunch of forks, simply using the path name can result in
the wrong project ID being found. This switches to using the
`path_with_namespace` value from the API to find the correct project.
Also adds a break to looping through the projects so that we do not
continue to search through the entire array when found.

Probably could look into using the search API endpoint, but as of yet
I've been unable to figure out exactly how to get the result expected.